Some additional thoughts about limitations
There are a couple of items that are listed as prerequisites that could also be considered limitations: 

* No stitching across report suites
* North America only

Should we list those in both places? Just under one or the other?

Also, I added a limitation relative to customer attributes. This is not related to VRS or RTP.  It's just a limitation of CDA.