Superscripts not always appearing properly
For Time Decay notes, in Chrome and Firefox current browsers 12/16/2019, the superscript representing 2^(-t/halflife) was not displaying properly. After working in the editor here in Chrome, it began to work, without having made any change. But the issue persists in Firefox and Chrome except in this session now. I suggest checking CSS interactions with <sup> tags or considering use of the carat symbol to ensure it is clear. Knowing what to look for, I was dumbfounded trying to interpret this appearance:

2 (-t/halflife)