Update apl.md
I've noticed that in the section "Install the plug-in using AppMeasurement" the actual code example is missing "s.apl =". If you copy this code and use it as is, it will break the tracking as the s.apl is not defined once you try to call the function. The solution above resolves this issue.