Update privacy-overview.md
The "privacy by design" seems to be outdated information and should be replaced with this existing setup that is available in the UI:

https://experienceleague.adobe.com/docs/analytics/admin/admin-tools/manage-report-suites/edit-report-suite/report-suite-general/general-acct-settings-admin.html?lang=en

Replace the last octet of IP addresses with 0	Removing the last octet is done before any processing is done on the hit, including before IP filtering/exclusion, before checking bot rules, before Geosegmentation lookups, etc. As such, the last octet is replaced with a 0, and IP exclusion rules should be updated to match IP addresses with a zero on the end. Matching * should match 0. For example, the IP address 11.22.33.44 gets changed to 11.22.33.0. Geosegmentation data will not be quite as exact as when the whole IP address is used. Specifically, city accuracy is going to be more affected than country or region accuracy. Both Bot rules and VISTA rules are affected because the entire IP address is unavailable to them. In addition, any processing rules that are IP based, including marketing channel rules and report suite processing rules, are affected by this setting.
Note: This setting is enabled by default for any new report suites created in the London Data Center after January 2019, but only if the settings for those report suites are copied from a template listed in the Admin Console. Report suites whose settings are duplicated from other report suites will inherit all settings from the selected report suite.